+++ test/system/admin_users_show_test.rb
@@ -32,7 +32,6 @@ class AdminUsersShowTest < ApplicationSystemTestCase
   end
 
   # rubocop:disable Minitest/MultipleAssertions
-  # TODO: our tests are quite smelly.  This one needs work!
   test 'should be able to toggle suspended/admin a regular user' do
     admin = users(:user_admin)
     user = users(:user_regular)
@@ -105,7 +104,6 @@ class AdminUsersShowTest < ApplicationSystemTestCase
 
     logout_user
   end
-  # rubocop:enable Minitest/MultipleAssertions
 
   test 'should be able to login as a regular user' do
     admin = users(:user_admin)
@@ -150,6 +148,7 @@ class AdminUsersShowTest < ApplicationSystemTestCase
 
     logout_user
   end
+  # rubocop:enable Minitest/MultipleAssertions
 
   test 'should be able to view items/theses owned by user' do
     # NOTE: searching and faceting is covered more extensively in tests elsewhere